refactor "patron hold"
