+++ src/main/java/org/folio/innreach/controller/d2ir/InnReachCirculationController.java
@@ -4,6 +4,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
@@ -12,6 +13,7 @@ import org.springframework.web.bind.annotation.RestController;
 import org.folio.innreach.domain.service.CirculationService;
 import org.folio.innreach.dto.CirculationRequestDTO;
 import org.folio.innreach.dto.InnReachResponseDTO;
+import org.folio.innreach.dto.PatronHoldDTO;
 import org.folio.innreach.dto.TransferRequestDTO;
 import org.folio.innreach.rest.resource.InnReachCirculationApi;
 
@@ -34,8 +36,17 @@ public class InnReachCirculationController implements InnReachCirculationApi {
   }
 
   @Override
+  @PostMapping("/patronHold/{trackingId}/{centralCode}")
+  public ResponseEntity<InnReachResponseDTO> patronHold(@PathVariable String trackingId,
+      @PathVariable String centralCode, PatronHoldDTO patronHold) {
+    var innReachResponse = circulationService.initiatePatronHold(trackingId, centralCode, patronHold);
+
+    return ResponseEntity.ok(innReachResponse);
+  }
+
+  @Override
   @PutMapping("/transferrequest/{trackingId}/{centralCode}")
-  public ResponseEntity<InnReachResponseDTO> putTransferRequest(@PathVariable String trackingId,
+  public ResponseEntity<InnReachResponseDTO> transferRequest(@PathVariable String trackingId,
       @PathVariable String centralCode, TransferRequestDTO transferRequest) {
     var innReachResponse = circulationService.transferItem(trackingId, centralCode, transferRequest);
