+++ src/main/java/org/folio/innreach/mapper/InnReachTransactionHoldMapper.java
@@ -6,7 +6,7 @@ import org.mapstruct.Mapper;
 import org.folio.innreach.domain.entity.TransactionItemHold;
 import org.folio.innreach.domain.entity.TransactionLocalHold;
 import org.folio.innreach.domain.entity.TransactionPatronHold;
-import org.folio.innreach.dto.CirculationRequestDTO;
+import org.folio.innreach.dto.PatronHoldDTO;
 import org.folio.innreach.dto.TransactionHoldDTO;
 
 @Mapper(componentModel = "spring", injectionStrategy = InjectionStrategy.CONSTRUCTOR, uses = {MappingMethods.class, InnReachTransactionPickupLocationMapper.class})
@@ -27,5 +27,5 @@ public interface InnReachTransactionHoldMapper {
 
   TransactionPatronHold toPatronHold(TransactionHoldDTO dto);
 
-  TransactionHoldDTO mapRequest(CirculationRequestDTO dto);
+  TransactionHoldDTO mapRequest(PatronHoldDTO dto);
 }
