+++ src/main/resources/swagger.api/circulation.yaml
@@ -169,9 +169,38 @@ paths:
         - $ref: '#/components/parameters/trackingId'
         - $ref: '#/components/parameters/centralCode'
         - $ref: '#/components/parameters/circulationOperationName'
+  /d2ir/circ/patronhold/{trackingId}/{centralCode}:
+    post:
+      operationId: patronHold
+      description: Originate a patron hold
+      tags:
+        - inn-reach-circulation
+      requestBody:
+        content:
+          application/json:
+            schema:
+              $ref: "#/components/schemas/patronHoldDTO"
+      responses:
+        '200':
+          description: Transfer request complete
+          content:
+            application/json:
+              schema:
+                $ref: "#/components/schemas/innReachResponseDTO"
+        '400':
+          description: An error occurred during processing the request
+          content:
+            application/json:
+              schema:
+                $ref: "#/components/schemas/innReachResponseDTO"
+        '500':
+          $ref: "api-common.yaml#/components/responses/trait_response_500"
+      parameters:
+        - $ref: '#/components/parameters/trackingId'
+        - $ref: '#/components/parameters/centralCode'
   /d2ir/circ/transferrequest/{trackingId}/{centralCode}:
     put:
-      operationId: putTransferRequest
+      operationId: transferRequest
       description: Put transfer request
       tags:
         - inn-reach-circulation
@@ -262,6 +291,8 @@ components:
       $ref: schemas/transactionHoldDTO.json
     circulationRequestDTO:
       $ref: schemas/d2ir/circulationRequestDTO.json
+    patronHoldDTO:
+      $ref: schemas/d2ir/patronHoldDTO.json
     transferRequestDTO:
       $ref: schemas/d2ir/transferRequestDTO.json
     innReachTransactionsDTO:
