+++ src/main/resources/swagger.api/schemas/d2ir/baseCircRequestDTO.json
