+++ src/main/resources/swagger.api/schemas/d2ir/patronHoldDTO.json
@@ -1,0 +1,46 @@
+{
+  "$schema": "http://json-schema.org/draft-04/schema#",
+  "allOf": [
+    {
+      "$ref": "baseCircRequestDTO.json"
+    }
+  ],
+  "description": "D2IR Patron hold",
+  "type": "object",
+  "properties": {
+    "pickupLocation": {
+      "description": "Pickup location",
+      "type": "string"
+    },
+    "centralItemType": {
+      "description": "Central item type",
+      "type": "integer",
+      "minimum": 0,
+      "maximum": 255
+    },
+    "title": {
+      "description": "Item title",
+      "type": "string",
+      "maxLength": 256
+    },
+    "author": {
+      "description": "Author",
+      "type": "string",
+      "maxLength": 256
+    },
+    "callNumber": {
+      "description": "Call number",
+      "type": "string",
+      "maxLength": 256
+    },
+    "needBefore": {
+      "description": "Need before time",
+      "type": "integer"
+    }
+  },
+  "additionalProperties": false,
+  "required": [
+    "pickupLocation",
+    "centralItemType"
+  ]
+}
